+++ pyincore/analyses/epfdamage/epfdamage.py
@@ -9,7 +9,7 @@ from itertools import repeat
 
 from pyincore import AnalysisUtil, GeoUtil
 from pyincore import BaseAnalysis, HazardService, FragilityService
-from pyincore.models.fragilitycurve import FragilityCurve
+from pyincore.models.dfr3curve import DFR3Curve
 
 
 class EpfDamage(BaseAnalysis):
@@ -134,8 +134,8 @@ class EpfDamage(BaseAnalysis):
                 geology_dataset_id = self.get_parameter("liquefaction_geology_dataset_id")
 
                 if geology_dataset_id is not None:
-                    fragility_sets_liq = self.fragilitysvc.match_inventory( self.get_input_dataset("dfr3_mapping_set"),
-                                                                            epfs, liquefaction_fragility_key)
+                    fragility_sets_liq = self.fragilitysvc.match_inventory(
+                        self.get_input_dataset("dfr3_mapping_set"), epfs, liquefaction_fragility_key)
 
                     if fragility_sets_liq is not None:
                         liquefaction_available = True
@@ -199,7 +199,7 @@ class EpfDamage(BaseAnalysis):
             damage_result = dict()
             selected_fragility_set = fragility_set[epf["id"]]
 
-            if isinstance(selected_fragility_set.fragility_curves[0], FragilityCurve):
+            if isinstance(selected_fragility_set.fragility_curves[0], DFR3Curve):
                 hazard_val = AnalysisUtil.update_precision_of_lists(hazard_vals[i]["hazardValues"])
                 input_demand_types = hazard_vals[i]["demands"]
                 input_demand_units = hazard_vals[i]["units"]
@@ -218,7 +218,7 @@ class EpfDamage(BaseAnalysis):
                 if liquefaction_resp is not None:
                     fragility_set_liq = fragility_sets_liq[epf["id"]]
 
-                    if isinstance(fragility_set_liq.fragility_curves[0], FragilityCurve):
+                    if isinstance(fragility_set_liq.fragility_curves[0], DFR3Curve):
                         liq_hazard_vals = AnalysisUtil.update_precision_of_lists(liquefaction_resp[i]["pgdValues"])
                         liq_demand_types = liquefaction_resp[i]["demands"]
                         liq_demand_units = liquefaction_resp[i]["units"]
@@ -240,9 +240,8 @@ class EpfDamage(BaseAnalysis):
 
                     limit_states = AnalysisUtil.adjust_limit_states_for_pgd(limit_states, pgd_limit_states)
 
-                dmg_interval = selected_fragility_set.calculate_damage_interval(limit_states,
-                                                                       hazard_type=hazard_type,
-                                                                       inventory_type='electric_facility')
+                dmg_interval = selected_fragility_set.calculate_damage_interval(
+                    limit_states, hazard_type=hazard_type, inventory_type='electric_facility')
             else:
                 raise ValueError("One of the fragilities is in deprecated format. This should not happen. If you are "
                                  "seeing this please report the issue.")