+++ pyincore/utils/analysisutil.py
@@ -68,7 +68,6 @@ class AnalysisUtil:
     def float_dict_to_decimal(num_dict: dict):
         return {key: Decimal(str(num_dict[key])) for key in num_dict}
 
-
     @staticmethod
     def calculate_mean_damage(dmg_ratio_tbl, dmg_intervals,
                               damage_interval_keys, is_bridge=False,
@@ -380,7 +379,7 @@ class AnalysisUtil:
 
             building_args = fragility_set.construct_expression_args_from_inventory(building)
             building_period = fragility_set.fragility_curves[0].get_building_period(
-                fragility_set.fragility_curve_parameters, **building_args)
+                fragility_set.curve_parameters, **building_args)
 
             if fragility_hazard_type.endswith('sa') and fragility_hazard_type != 'sa':
                 # This fixes a bug where demand type is in a format similar to 1.0 Sec Sa
@@ -444,7 +443,7 @@ class AnalysisUtil:
                 # Get building period from the fragility if possible
                 building_args = fragility_set.construct_expression_args_from_inventory(building)
                 building_period = fragility_set.fragility_curves[0].get_building_period(
-                    fragility_set.fragility_curve_parameters, **building_args)
+                    fragility_set.curve_parameters, **building_args)
 
                 # TODO: There might be a bug here as this is not handling SD
                 if demand_type.endswith('sa'):
@@ -460,7 +459,7 @@ class AnalysisUtil:
                                 building_period = building[PROPERTIES][BLDG_PERIOD]
                             else:
                                 # try to calculate the period from the expression
-                                for param in fragility_set.fragility_curve_parameters:
+                                for param in fragility_set.curve_parameters:
                                     if param["name"].lower() == "period":
                                         # TODO: This is a hack and expects a parameter with name "period" present.
                                         #  This can potentially cause naming conflicts in some fragilities